[refact] Global engine.py refactor
All changes in this commit are aimed at optimizing the game.

Two new project dependencies have been added:

    - numpy - for using efficient arrays.

    - numba - for fast code compilation.

Update type hints.
Rename method next_cycle() -> process()
Added the main function of calculating the next step in the game -
 - check_cell()

Added new logic for the states of the playing field:

    * current_area - The current state of game area

    * next_field - The next state of game area

    * draw_rects - The rects that will be draw on game area

Completely abandoned the use of the Cell class in favor
of the performance of innovations.