[feature] Improve performance by add function
Added cell neighbors counting function - count_neighbors()

Now the responsibility for counting 8 cell neighbors lies with
count_neighbors()

The function does not count in a tricky way: instead of iterating over
all 8 neighbors, it counts only 4. The count is done in pairs in a 3x3
field.

For example, if the first cell is considered, then 9 (Mirror of
the center) is also considered, etc.

I also fixed a bug that allowed cells to survive on the borders,
because the previous version of the check_cells() function checked not
only the boundary values of the array, but also went beyond them,
thereby entering from the other side...

Now this is fixed and the area that is displayed on the screen is
calculated.

But there are minor remarks:

    * Confusion with x and y regarding actual cell placement.

    * Confusion with field width and height being interpreted
        differently.

Changes:
    change:     src/engine.py