[feature] Force FPS reduction on pause
Since a static image is displayed during a pause, it makes no sense to
allocate N frames for this, because It makes sense to show N frames
with a dynamic picture, but having a static one is pointless.

Thus, the load on the processor during the pause is released.

Changes:
    change:[update]     src/app.py