[feat] Add rules for game
An enumeration of "rules" was implemented, which is a special string
of the form "bN/sT" where N means how many living neighbors are
needed in order for a new cell to appear, and T is how many are needed
for the cell to remain alive.

Properties have also been added to `GameEngine` to change the rules that
are parsed in the `process` method and passed as two tuples to the
`check_cells` function.

This innovation will allow you to freely manipulate the rules for cells
in the game.

Files:
    [ch] base.py
    [ch] engine.py